requirements.txt
anthropic>=0.40.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
tqdm==4.66.4
scikit-learn==1.5.0
pillow==10.3.0
httpx==0.27.0
python-dotenv==1.0.1
